+++ CHANGES.md
@@ -12,6 +12,7 @@
 - Slack/Markdown: Fixed encoding link titles including square brackets
 - Time Interval: Fixed `week_to_day_range`
 - GitHub/Actions: Permitted commandeering the time interval per `--when` option
+- GitHub/Actions: Added filtering duplicates
 
 ## v0.3.0, 2025-02-24
 - Opsgenie: Added alert reporter. Thanks, @WalBah.